Updated character selction to work with main files
