+++ package.json
@@ -9,10 +9,10 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.20.7",
-    "@types/socket.io": "^3.0.1",
-    "@types/socket.io-client": "^1.4.36",
     "@tailwindcss/postcss": "^4.1.5",
     "@tailwindcss/vite": "^4.1.6",
+    "@types/socket.io": "^3.0.1",
+    "@types/socket.io-client": "^1.4.36",
     "meteor-node-stubs": "^1.2.5",
     "postcss": "^8.5.3",
     "postcss-load-config": "^6.0.1",
